elk: elasticsearch ingest node rules to merge multi log lines
This example is very helpful to verfy the solution that filebeat work
as the shipper and elasticsearch play as the indexer/search/storage.

Beside this commit is tring to solve the problem that merge the orch
multi-lines of exception logs into signel field.

Original log format like this:
  {"log": "This is KK"}
  {"log": "This is FX"}

I wish to merge the json field of 'log' like this:
  {"log": "This is KK\nThis is FX"}

Signed-off-by: Kuankuan Yang <kkyang@astri.org>